+++ CHANGELOG.md
@@ -10,6 +10,7 @@
 - PR #867 CSV Reader: Add support for ignoring blank lines and comment lines
 - PR #895 Add Series groupby
 - PR #898 Add DataFrame.groupby(level=0) support
+- PR #888 CSV Reader: Add prefix parameter for column names, used when parsing without a header
 
 ## Improvements